# api/openapi.yaml
openapi: 3.1.0
info:
  title: Nightjar API
  version: 0.2.0
  description: |
    Public API contract. Server handlers and the web client are built against
    this file. v0: libraries, items, playback info, direct-play and remux
    streams (ADR-0006).
  license:
    name: GPL-3.0-only
servers:
  - url: http://localhost:8096
    description: Local default
paths:
...
  /api/v0/items/{itemId}/subtitles/{trackId}.vtt:
    get:
      operationId: getSubtitleVtt
      summary: WebVTT for one text subtitle track (ADR-0010)
      parameters:
        - $ref: "#/components/parameters/ItemId"
        - name: trackId
          in: path
          required: true
          description: >
            Stable track id (ADR-0010): `e{streamIndex}` for embedded,
            `s` / `s-{suffix}` for filesystem sidecars.
          schema:
            type: string
            pattern: '^[es]([A-Za-z0-9._-]*)$'
      responses:
        "200":
          description: WebVTT body
          content:
            text/vtt:
              schema:
                type: string
        "404":
          description: Item or serveable subtitle track not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/v0/items/{itemId}/stream:
    get:
      operationId: streamItem
      summary: Byte stream for playback (supports HTTP Range)
      description: >
        Serves the original file for directPlay items and the remuxed MP4 for
        remux items once remuxState is ready.
      parameters:
        - $ref: "#/components/parameters/ItemId"
        - name: Range
          in: header
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Entire file
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "206":
          description: Partial content
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Remux not ready; POST remux and poll playback-info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "415":
          description: Item needs HLS transcode; POST /api/v0/items/{itemId}/sessions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "416":
          description: Range not satisfiable
...
components:
  parameters:
...
    ItemId:
      name: itemId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...